+++ R/oda_get_crs.R
@@ -1,0 +1,12 @@
+oda_get_dac1 <- function(
+  start_year,
+  end_year,
+  filters,
+  pre_process,
+  dotstat_codes,
+  dataflow_version,
+  as_grant_equivalent
+) {
+  req <- create_request(resource = "DSD_CRS@DF_CRS")
+  resp <- perform_request(req, "oda_get_crs")
+}