+++ R/oda_get_multisystem.R
@@ -1,0 +1,11 @@
+oda_get_multisystem <- function(
+  start_year,
+  end_year,
+  filters,
+  pre_process,
+  dotstat_codes,
+  dataflow_version
+) {
+  req <- create_request(resource = "DSD_MULTI@DF_MULTI")
+  resp <- perform_request(req, "oda_get_multisystem")
+}