+++ R/oecdoda-package.R
@@ -2,9 +2,9 @@
 #' @import cli
 #' @import httr2
 #' @import tibble
-#' @import jsonlite
 "_PACKAGE"
 
 ## usethis namespace: start
+#' @importFrom utils read.csv
 ## usethis namespace: end
 NULL