+++ NEWS.md
@@ -1,3 +1,3 @@
 # oecdoda (development version)
 
-* Initial CRAN submission.
+* Initial CRAN submission with `oda_get_dac1()`